# abstract the parquet coalescing reading