--- a/sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuMultiFileReader.scala
+++ b/sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuMultiFileReader.scala
@@ -185,7 +185,7 @@ abstract class MultiFileCloudPartitionReaderBase(
    * 2. Different file formats have different ThreadPoolExecutors
    *
    * @param numThreads  max number of threads to create
-   * @return ThreadPoolExecutors
+   * @return ThreadPoolExecutor
    */
   def getThreadPool(numThreads: Int): ThreadPoolExecutor
 
@@ -338,24 +338,29 @@ trait SingleDataBlockInfo {
 
 /**
  * The abstracted multi-file coalescing reading class, which tries to coalesce small
- * ColumnarBatches into a bigger ColumnarBatch according to maxReadBatchSizeRows,
+ * ColumnarBatch into a bigger ColumnarBatch according to maxReadBatchSizeRows,
  * maxReadBatchSizeBytes and others.
  *
- * The whole data reading process will be typically benefit from bigger ColumnarBatch.
+ * Please be note, this class is applied to below similar file format
+ *
+ * | HEADER |   -> optional
+ * | block  |   -> repeated
+ * | FOOTER |   -> optional
  *
  * The data driven:
  *
- * next() -> populateCurrentBlockChunk (try the best to coalesce ColumnarBatches)
+ * next() -> populateCurrentBlockChunk (try the best to coalesce ColumnarBatch)
  *        -> allocate a bigger HostMemoryBuffer for the populated block chunks
  *        -> writer header
- *        -> launch threads (each thread for each block) to copy the blocks to the HostMemoryBuffer
+ *        -> launch tasks to copy the blocks to the HostMemoryBuffer
+ *        -> wait all tasks finished
  *        -> writer footer
  *        -> decode the HostMemoryBuffer in the GPU
  *
- * @param clippedBlocks         the block metadata from the original Parquet file that has been
+ * @param clippedBlocks         the block metadata from the original file that has been
  *                                clipped to only contain the column chunks to be read
  * @param readDataSchema        the Spark schema describing what will be read
- * @param partitionSchema       Schema of partitions
+ * @param partitionSchema       schema of partitions
  * @param maxReadBatchSizeRows  soft limit on the maximum number of rows the reader reads per batch
  * @param maxReadBatchSizeBytes soft limit on the maximum number of bytes the reader reads per batch
  * @param numThreads            the size of the threadpool
@@ -393,7 +398,7 @@ abstract class MultiFileCoalescingPartitionReaderBase(
    *
    * @param currentBlockInfo current SingleDataBlockInfo
    * @param nextBlockInfo    next SingleDataBlockInfo
-   * @return Boolean
+   * @return true: split the next block into another ColumnarBatch and vice versa
    */
   def checkIfNeededToSplitDataBlock(
     currentBlockInfo: SingleDataBlockInfo,
@@ -427,11 +432,14 @@ abstract class MultiFileCoalescingPartitionReaderBase(
    * which will be running in a thread pool
    *
    * @param file   file to be read
-   * @param outhmb the sliced HostMemoryBuffer to hold the blocks,
-   *               which should be closed in sub-class
-   * @param blocks blocks meta info to specify which block to be copied
+   * @param outhmb the sliced HostMemoryBuffer to hold the blocks, and the user is
+   *               in charge of closing it in sub-class
+   * @param blocks blocks meta info to specify which blocks to be read
    * @param offset used as the offset adjustment
-   * @return Callable[(Seq[DataBlockBase], Long)]
+   * @return Callable[(Seq[DataBlockBase], Long)], which will be submitted to a
+   *         ThreadPoolExecutor, and the Callable will return a tuple result and
+   *         result._1 is block meta info with the offset adjusted
+   *         result._2 is the bytes read
    */
   def getBatchRunner(
     file: Path,
@@ -450,11 +458,9 @@ abstract class MultiFileCoalescingPartitionReaderBase(
   /**
    * Sent host memory to GPU to decode
    *
-   * Please be note, the dataBuffer only contains the data blocks, not including magic and footer
-   *
-   * @param dataBuffer  data block in Host Memory
+   * @param dataBuffer  the data which can be decoded in GPU
    * @param dataSize    data size
-   * @param isCorrectRebaseMode specify if need to rebase
+   * @param isCorrectRebaseMode rebase mode
    * @param clippedSchema the clipped schema
    * @return Table
    */
@@ -478,14 +484,14 @@ abstract class MultiFileCoalescingPartitionReaderBase(
    * estimated initialized buffer size may be a little smaller than the actual size. So in
    * this case, the hmb should be closed in the implementation.
    *
-   * @param hmb            The buffer holding (header + data blocks)
+   * @param hmb            The original buffer holding (header + data blocks)
    * @param initTotalSize  The initialized buffer size
-   * @param offset         Where begin to write
+   * @param footerOffset   Where begin to write the footer
    * @param blocks         The data block meta info
    * @param clippedSchema  The clipped schema info
    * @return the final HostMemoryBuffer and its size
    */
-  def writeFileFooter(hmb: HostMemoryBuffer, initTotalSize: Long, offset: Long,
+  def writeFileFooter(hmb: HostMemoryBuffer, initTotalSize: Long, footerOffset: Long,
     blocks: Seq[DataBlockBase], clippedSchema: SchemaBase): (HostMemoryBuffer, Long)
 
   override def get(): ColumnarBatch = {
@@ -518,7 +524,7 @@ abstract class MultiFileCoalescingPartitionReaderBase(
   }
 
   private def readBatch(): Option[ColumnarBatch] = {
-    withResource(new NvtxRange("Parquet readBatch", NvtxColor.GREEN)) { _ =>
+    withResource(new NvtxRange(s"$getFileFormatShortName readBatch", NvtxColor.GREEN)) { _ =>
       val currentChunkMeta = populateCurrentBlockChunk()
       if (readDataSchema.isEmpty) {
         // not reading any data, so return a degenerate ColumnarBatch with the row count
@@ -599,7 +605,7 @@ abstract class MultiFileCoalescingPartitionReaderBase(
 
       val allBlocks = blocks.map(_._2)
 
-      // First, estimate the output file size for further allocating a buffer for it,
+      // First, estimate the output file size for the initial allocating
       val initTotalSize = calculateEstimatedBlocksOutputSize(allBlocks, clippedSchema)
 
       val (buf, offset, outBlocks) = closeOnExcept(HostMemoryBuffer.allocate(initTotalSize)) {