@tgravescs 2021-06-17
Also what testing have you done here?  I would like to run a few performance tests to make sure no regressions, if we can run on a bunch of small files and compares current version vs this version that would be great.