@tgravescs 2021-06-18
fyi - still 2 questions/comments open