@wbo4958 2021-06-21
Hi @tgravescs 

I've tested the below scenarios for `upstream` and `upstream + this PR`

- small file testing

total 4000 small parquet files, total 16M

||upstream + this PR time(s)| upstream time(s)|
|---|----|---|
| 1st |  35.736 |33.976|
|2nd|27.406|27.346|
|3rd|23.98|23.442|
|4th|26.217|26.331|

- a litter bigger file testing

total 100 parquet files,  each file is 14M, total 1.3G

||upstream + this PR, time(s)| upstream time(s) |
|---|----|---|
| 1st |  13.836  | 13.421|
|2nd|12.348|12.289|
|3rd|11.676|11.175|
|4th|10.966|10.842|

Seems this PR does not bring performance regression.